backend/requirements.txt
fastapi==0.95.2
uvicorn[standard]==0.22.0
pydantic==2.6.2
python-dotenv==1.0.0
asyncpg==0.27.0
google-genai==1.52.0
